backend/requirements.txt
fastapi
uvicorn
python-multipart
transformers
torch
pillow
opencv-python

PyWavelets
scikit-image
scipy
ftfy
regex
mediapipe
